// typesnipe-ui/src/Api.ts
export interface DeviceOrientationModel {
  /** Alpha */
  alpha: number;
  /** Beta */
  beta: number;
  /** Gamma */
  gamma: number;
}

/** HTTPValidationError */
export interface HTTPValidationError {
  /** Detail */
  detail?: ValidationError[];
}

/** KeystrokeModel */
export interface KeystrokeModel {
  /** Value */
  value: string;
  orientation: DeviceOrientationModel;
  /** Time */
  time: number;
}